pyproject.toml
[project]
name = "slack-chat"
version = "2.0.0"
description = "Slack CLI via browser-use"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use",
    "fastapi",
    "uvicorn",
    "typer",
    "httpx",
    "python-dotenv",
    "pyyaml",
]

[project.scripts]
slack-chat = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
